THOR-1361 Fixed CustomRequestsComponent and renamed parent folder.  Changed to navbar hamburger menu item.  Can now show multiple custom requests defined in config file.  Other minor code cleanup.
