+++ src/app/dashboard/dashboard.component.ts
@@ -92,7 +92,6 @@ export class DashboardComponent implements AfterViewInit, OnInit, OnDestroy {
     // Toolbar
     showVisualizationsShortcut: boolean = true;
     showDashboardSelector: boolean = false;
-    showCustomRequests: boolean = false;
 
     rightPanelTitle: string;
 
@@ -503,19 +502,11 @@ export class DashboardComponent implements AfterViewInit, OnInit, OnDestroy {
     toggleFiltersDialog() {
         this.showFiltersComponent = !this.showFiltersComponent;
         this.showDashboardSelector = false;
-        this.showCustomRequests = false;
-    }
-
-    toggleCustomRequests() {
-        this.showCustomRequests = !this.showCustomRequests;
-        this.showFiltersComponent = false;
-        this.showDashboardSelector = false;
     }
 
     toggleDashboardSelectorDialog() {
         this.showDashboardSelector = !this.showDashboardSelector;
         this.showFiltersComponent = false;
-        this.showCustomRequests = false;
     }
 
     /**
